aj07mm-api-test/requirements.txt
aniso8601==1.3.0
attrs==17.4.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
factory-boy==2.9.2
Faker==0.8.8
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
MarkupSafe==1.0
marshmallow==2.15.0
mongoengine==0.15.0
nose==1.3.7
pluggy==0.6.0
py==1.5.2
PyJWT==1.5.3
pymongo==3.6.0
pytest==3.3.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
text-unidecode==1.1
urllib3==1.22
Werkzeug==0.13
